pyproject.toml
[project]
name = "academy-first-rag-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "datasets>=4.0.0",
    "fastapi>=0.116.1",
    "h5py>=3.14.0",
    "numpy>=2.2.6",
    "pandas>=2.3.1",
    "pyarrow>=21.0.0",
    "python-dotenv>=1.1.1",
    "tqdm>=4.67.1",
    "weaviate-client>=4.16.6",
]
